\name{bioanthdatasets-package}
\alias{bioanthdatasets-package}
\alias{bioanthdatasets}
\docType{package}
\title{
\packageTitle{bioanthdatasets}
}
\description{
\packageDescription{bioanthdatasets}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bioanthdatasets}
\packageIndices{bioanthdatasets}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bioanthdatasets}

Maintainer: \packageMaintainer{bioanthdatasets}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
