% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enamelthickness}
\alias{enamelthickness}
\title{Enamel thickness and other molar data for hominoids.}
\format{A data frame with 195 rows and 9 variables: 
\describe{ 
  \item{accession}{specimen identifier} 
  \item{taxon}{hominoid species} 
  \item{tooth}{mandibular molar M1, M2, or M3} 
  \item{enamel_area}{enamel area, in mm squared} 
  \item{dentine_area}{dentine area, in mm squared} 
  \item{edj_length}{enamel-dentine junction length, in mm} 
  \item{bcd}{bi-cervical diameter, in mm} 
  \item{aet}{average enamel thickness, in mm} 
  \item{ret}{relative enamel thickness} 
}}
\source{
Skinner MM, Alemseged Z, Gaunitz C, Hublin J-J. 2015. Enamel
  thickness trends in Plio-Pleistocene hominin mandibular molars.
  \emph{Journal of Human Evolution} 85:35-45. 
  \url{http://dx.doi.org/10.1016/j.jhevol.2015.03.012}
  
  The data were extracted from the supplementary material. The following
  modifications were made:
  \itemize{
    \item citation and basis for tooth identification columns dropped \item
    consolidated LM1 and RM1 into just M1, and the same for M2 and M3; none
    of the analyses in the paper itself differentiate between right and left 
    \item fixed an apparent error in the values for bcd, aet, and ret for the
    specimens of unknown taxonomic attribution (column values were rotated)
  }
}
\usage{
\code{data(enamelthickness)}
}
\description{
A dataset containing enamel thickness, enamel area, and other data (see
below) for 195 specimens representing 13 species of hominoids.
}
\examples{
boxplot(aet ~ taxon, data=enamelthickness, 
                     las=2, cex.axis=0.65)
}
\keyword{datasets}
