% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energyexpenditure}
\alias{energyexpenditure}
\title{Total energy expenditure and mass for mammals.}
\format{A data frame with 86 rows and 5 variables: 
\describe{ 
  \item{species}{species name} 
  \item{order}{mammalian order} 
  \item{captive}{yes, no, or unknown} 
  \item{mass}{in kg} 
  \item{tee}{total energy expenditure, in kcalories/day} 
}}
\source{
Pontzer H, Raichlin DA, Gordon AD, Schroepfer-Walker KK, Hare B, 
  O'Neill MC, Muldoon KM, Dunsworth HM, Wood BM, Isler K, Burkart J, Irwin M,
  Shumaker RW, Lonsdorf EV, Ross SR. 2014. Primate energy expenditure and
  life history. \emph{Proc Natl Acad Sci USA} 111:1433-1437. 
  \url{http://dx.doi.org/10.1073/pnas.1316940111}
  
  The data were extracted from the supplementary material. The following
  modifications were made:
  \itemize{
    \item renamed the admixed \emph{Eulemur sp.} as \emph{Eulemur}
    \item dropped the second \emph{Peromyscus maniculatus} data point (which
    was not used in primate - nonprimate comparisons in Pontzer et al.)
    \item \emph{Eremitalpa namibensis} is not a rodent, it's an Afrosoricid 
  }
}
\usage{
\code{data(energyexpenditure)}
}
\description{
A dataset containing total daily energy expenditure and mass for 86 species
of mammals.
}
\examples{
boxplot(aet ~ taxon, data=enamelthickness, 
                     las=2, cex.axis=0.65)
}
\seealso{
\code{\link{energyexpendituretree}} - a matching newick format
  tree for phylogenetic regression.
}
\keyword{datasets}
