% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{energyexpendituretree}
\alias{energyexpendituretree}
\title{Phylogenetic tree for 81 mammalian species.}
\format{A phylogenetic tree in the \code{ape} \code{\link[ape]{phylo}}
  format.}
\source{
Bininda-Emonds ORP, Cardillo M, Jone KE, MacPhee RDE, Deck RMD,
  Grenyer R, Price SA, Vos RA, Gittleman JL, Purvis A. The delayed rise of
  present-day mammals. \emph{Proc Natl Acad Sci USA} 111:1433-1437. 
  \url{http://dx.doi.org/10.1038/nature05634}
  
  The tree was taken from the supplementary material. The following
  modifications were made:
  \itemize{
    \item reduced to the taxa present in the \code{\link{energyexpenditure}}
    dataset.
    \item \emph{Euleumur fulvus} used as the tip for the admixed
    \emph{Eulemur} group
    \item \emph{Proteles cristatus} renamed to \emph{Proteles cristata} (typo
    in source)
    \item \emph{Homo sapiens} renamed to \emph{Homo sapiens} Hadza
    \item \emph{Canis rufus} used for \emph{Canis familiaris}
    \item \emph{Chrysochloris asiatica} used for \emph{Eremitalpa namibensis}
    \item \emph{Lemmus sibiricus} used for \emph{Lemmus trimucronatus}
  }
}
\usage{
\code{data(energyexpendituretree)}
}
\description{
A phylogenetic tree for 81 mammalian species in the 
\code{\link{energyexpenditure}} dataset.
}
\examples{
plot(energyexpendituretree, cex=0.5)
}
\seealso{
\code{\link{energyexpenditure}} - a matching dataset of mammalian
  mass and total energy expenditure.
}
\keyword{datasets}
